IOS-1943 Removed userwalletlistservice
