+++ Tangem/App/Services/UserWalletRepository/UserWalletRepository.swift
@@ -11,16 +11,29 @@ import Combine
 
 protocol UserWalletRepository {
     var delegate: UserWalletRepositoryDelegate? { get set }
-
     var models: [CardViewModel] { get }
+    var selectedModel: CardViewModel? { get }
+    var selectedUserWalletId: Data? { get set }
+    var isEmpty: Bool { get }
+    var isUnlocked: Bool { get }
 
     func scan(with batch: String?, requestBiometrics: Bool, _ completion: @escaping (Result<CardViewModel, Error>) -> Void)
     func scanPublisher(with batch: String?, requestBiometrics: Bool) ->  AnyPublisher<CardViewModel, Error>
 
     func add(_ cardModels: [CardViewModel])
     func removeModel(with userWalletId: Data)
-    func clear()
     func didSwitch(to cardModel: CardViewModel)
+
+    func lock()
+    func unlockWithBiometry(completion: @escaping (Result<Void, Error>) -> Void)
+    func unlockWithCard(_ userWallet: UserWallet, completion: @escaping (Result<Void, Error>) -> Void)
+
+    func didScan(card: CardDTO)
+
+    func contains(_ userWallet: UserWallet) -> Bool
+    func save(_ userWallet: UserWallet)
+    func delete(_ userWallet: UserWallet)
+    func clear()
 }
 
 protocol UserWalletRepositoryDelegate: AnyObject {
